Implementará Middleware
